# specs/001-guest-lock-pins/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Guest Lock PIN Manager API
  description: |
    REST API for the Guest Lock PIN Manager Home Assistant addon.
    Manages calendar subscriptions, lock configurations, and PIN lifecycle.
  version: 1.0.0
  contact:
    name: Guest Lock Manager
servers:
  - url: /api
    description: Home Assistant ingress-proxied API
...
paths:
...
  /locks/{id}:
    parameters:
      - $ref: '#/components/parameters/LockId'
    get:
      tags: [locks]
      summary: Get lock details
      operationId: getLock
      responses:
        '200':
          description: Lock details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Lock'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [locks]
      summary: Update lock configuration
      operationId: updateLock
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LockUpdate'
      responses:
        '200':
          description: Lock updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Lock'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [locks]
      summary: Remove lock from management
      operationId: deleteLock
      responses:
        '204':
          description: Lock removed
        '404':
          $ref: '#/components/responses/NotFound'
  /locks/{id}/pins:
    parameters:
      - $ref: '#/components/parameters/LockId'
    get:
      tags: [locks]
      summary: Get all PINs on a lock
      operationId: getLockPins
      responses:
        '200':
          description: PINs on lock
          content:
            application/json:
              schema:
                type: object
                properties:
                  guest_pins:
                    type: array
                    items:
                      $ref: '#/components/schemas/GuestPinSummary'
                  static_pins:
                    type: array
                    items:
                      $ref: '#/components/schemas/StaticPinSummary'
                  available_slots:
                    type: integer

  # ============== GUEST PINS ==============
  /guest-pins:
    get:
      tags: [guest-pins]
      summary: List all guest PINs
      operationId: listGuestPins
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, active, expired, conflict]
        - name: calendar_id
          in: query
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of guest PINs
          content:
            application/json:
              schema:
                type: object
                properties:
                  pins:
                    type: array
                    items:
                      $ref: '#/components/schemas/GuestPin'
  /guest-pins/{id}:
    parameters:
      - $ref: '#/components/parameters/GuestPinId'
    get:
      tags: [guest-pins]
      summary: Get guest PIN details
      operationId: getGuestPin
      responses:
        '200':
          description: Guest PIN details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GuestPin'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [guest-pins]
      summary: Update guest PIN (set custom code)
      operationId: updateGuestPin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                custom_pin:
                  type: string
                  pattern: '^\d{4,8}$'
                  description: Owner-specified PIN override
      responses:
        '200':
          description: PIN updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GuestPin'
        '404':
          $ref: '#/components/responses/NotFound'
  /guest-pins/{id}/regenerate:
    parameters:
      - $ref: '#/components/parameters/GuestPinId'
    post:
      tags: [guest-pins]
      summary: Regenerate PIN code
      operationId: regenerateGuestPin
      responses:
        '200':
          description: PIN regenerated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GuestPin'
        '400':
          description: PIN not eligible for regeneration (check-in within 24h)
        '404':
          $ref: '#/components/responses/NotFound'

  # ============== STATIC PINS ==============
...
components:
  parameters:
...
    LockId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    GuestPinId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
          example: "not_found"
        message:
          type: string
          example: "Calendar not found"
        details:
          type: object
          additionalProperties: true
...
    Lock:
      type: object
      properties:
        id:
          type: string
          format: uuid
        entity_id:
          type: string
        name:
          type: string
        protocol:
          type: string
          enum: [zwave, zigbee, wifi, unknown]
        total_slots:
          type: integer
        guest_slots:
          type: integer
        static_slots:
          type: integer
        available_slots:
          type: integer
        online:
          type: boolean
        battery_level:
          type: integer
          nullable: true
        direct_integration:
          type: string
          enum: [zwave_js_ui, zigbee2mqtt]
          nullable: true
        last_seen_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
...
    LockUpdate:
      type: object
      properties:
        name:
          type: string
        guest_slots:
          type: integer
          minimum: 0
        static_slots:
          type: integer
          minimum: 0
...
    GuestPin:
      type: object
      properties:
        id:
          type: string
          format: uuid
        calendar_id:
          type: string
          format: uuid
        calendar_name:
          type: string
        event_summary:
          type: string
        pin_code:
          type: string
        generation_method:
          type: string
          enum: [custom, phone_last4, description_random, date_based]
        valid_from:
          type: string
          format: date-time
        valid_until:
          type: string
          format: date-time
        status:
          type: string
          enum: [pending, active, expired, conflict]
        regeneration_eligible:
          type: boolean
        locks:
          type: array
          items:
            $ref: '#/components/schemas/PinLockStatus'
        created_at:
          type: string
          format: date-time
    GuestPinSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        event_summary:
          type: string
        pin_code:
          type: string
        status:
          type: string
        valid_until:
          type: string
          format: date-time
    PinLockStatus:
      type: object
      properties:
        lock_id:
          type: string
          format: uuid
        lock_name:
          type: string
        slot_number:
          type: integer
        sync_status:
          type: string
          enum: [pending, synced, failed, removed]
...
    StaticPinSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        enabled:
          type: boolean
        currently_active:
          type: boolean
...